azure-functions/requirements.txt
azure-functions
azure-identity>=1.15.0
azure-storage-blob>=12.18.0
azure-data-tables>=12.4.2
azure-mgmt-resource>=23.0.0
azure-mgmt-web>=8.0.0
finnhub-python
requests
python-dotenv
pydantic
openai
pyyaml
numpy
pandas
sec-api
bcrypt
PyMuPDF
